split nanopub and move to whyis
