+++ whyis/nanopub/__init__.py
@@ -1,0 +1,2 @@
+from .nanopublication import Nanopublication
+from .nanopublication_manager import NanopublicationManager
